move controller package to bot package
