+++ bot/src/main/java/edu/java/bot/controller/exception/BadRequestException.java
@@ -1,4 +1,4 @@
-package edu.java.controller.exception;
+package edu.java.bot.controller.exception;
 
 public class BadRequestException extends RuntimeException {
     public BadRequestException(String message) {
